Incorporation of Mathieu's debugging
Incorporation in my code of Mathieu's debugging based on Winmerge comparison between 2 IAM_1.4.cpp files: the one from October 2018 and his debugged version.
It included modifications to the aggregObj function and calls to the function
He also checked the PROTECT/UNPROTECT but I haven't included those in my version as I already had treated the issue. However if my version of the model crashes when running it sequentially, it ;ight be worth looking deeper into it as Mathieu's modifications enabled to avoid the crashes we had before.
More details about the changes I included in my notebook nb. 2