Include exogenous quota prices
Allow for specification of exogenous quota prices, i.e quota prices for species for which we don't explicitely model the quota market (e.g. because they are not explicitely under quota management in the model).
When the quota mrket module is activated, it models the quota market of all species under TAC management (either with fixed TAC or under a HCR). All other species included in the input parOptQuot$pxQuIni will have an exogenous quota price.
NB: This is useful to avoid unrealistic effort shifts (when fishing behavious module is active)  towards unregulated species in the model.