QuotaMarket with PriceFlex
Make modifications for QuotaMarket to be functionnal with a Market module (fish price) active.
Originally, vessels would estimate their ProfPUE on the reference year (the easiest way since all vessels are active at t=0; otherwise vessels that are inactive in a competitive fishery (ie once quota market is on) have no means to estimate their ProfPUE.. --> In a future version, maybe try have everyone fish for fake in the current year and base ProfPUE on this.
What is done here to get around the problem is to adjust GVLUE from t=0 based on the ratio of price between t=1 and t=0

Also added a distinction between quota holders and vessels: now there is a possibility for external quota holders. External investors are assumed to have priority in leasing out quota compared to fishers