Initial commit for IAM 2.0.0
Added the source code from Florence, but I did not run the code on an
example.
NEWS.md file for tracking modification of the package
Licence change from GPL-2 to CeCILL.

Description need further information.

Package goes in check and return warning and notes, but it ends well
Package install and restart.