Init doc for 0_input.R
Added roxygen format for most classes, but need to work on 1_input
to understand format.
Once both files are done, need to work on example files to add to the
package.
And redact a vignette about this first part.

Need to understand the origin of output stockInput class, as it is not
used.

All empty validity functions are commented since they are not mandatory.

Maxime Jaunatre