NAMESPACE and correct dep
NAMESPACE is now correclty build by Roxygen.
To do so, all dependencies are now inside the doc, even if repeated for
each function.
All Mathieu'man is inside Roxygen format now.

Removed MASS dependence, because it was used to load a lot of minor dep.
Moved all dep from Depends to Imports, since it's a better practice !