Generic markdown and figure
Add simpliest way to use the package in main vignette.
However C++ part crash when running the code, I may be too brutal.

Use Notice to draw an understanding of the main functions interactions.
Not sure about the end of the pipeline, need more precisions. Also need
to have alternative way from the guy for selecting the args !