New IAM.format for extracting dataset+shinyplot
Remove old IAM.format and replace it with custom function.
Produce a new S3 table based on tibble for easy filtering.
Added a new shiny app for plot exploration, need to be tested a bit more

A lot of edits, but this is like a feature branch for 2week work.

Maxime Jaunatre